requirements.txt
GitPython
openai
PyYAML
python-dotenv
schedule
pytz